requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
streamlit
requests